projects/poem-generator/requirements.txt
flask
gunicorn
virtualenv
google-cloud-vision
lyricsgenius
google-cloud-texttospeech
google-cloud